Refactor registerProxyRoutes and add test
