+++ src/server/app.ts
@@ -3,13 +3,14 @@ import morgan from "morgan";
 import cors from "cors";
 import basicAuth from "express-basic-auth";
 import swaggerUi from "swagger-ui-express";
-import generalError, { unknownEndpoint } from "./middlewares/errors.js";
-import { paths } from "./routers/paths.js";
+import generalError, { unknownEndpoint } from "./middlewares/errors/errors.js";
 import pingPongProtocolRouter from "./routers/pingPongProtocolRouter/pingPongProtocolRouter.js";
 import openApiDocument from "../openapi/index.js";
 import corsOptions from "./cors/corsOptions.js";
 import { environment } from "../loadEnvironments.js";
-import { registerProxyRoutes } from "./middlewares/proxy.js";
+import { registerProxyRoutes } from "./middlewares/proxy/proxy.js";
+import proxyRoutes from "./routers/proxy/proxyRoutes.js";
+import { paths } from "./routers/paths.js";
 
 const app = express();
 
@@ -22,7 +23,7 @@ app.use(express.json());
 
 app.use(paths.root, pingPongProtocolRouter);
 
-registerProxyRoutes(app);
+registerProxyRoutes(app, proxyRoutes);
 
 app.use(
   paths.apiDocs.base,
