+++ src/server/middlewares/errors/errors.ts
@@ -1,16 +1,16 @@
-import "../../loadEnvironments.js";
+import "../../../loadEnvironments.js";
 import debugCreator from "debug";
 import chalk from "chalk";
 import type { NextFunction, Request, Response } from "express";
-import CustomError from "../../CustomError/CustomError.js";
-import httpStatusCodes from "../../utils/httpStatusCodes.js";
+import CustomError from "../../../CustomError/CustomError.js";
+import httpStatusCodes from "../../../utils/httpStatusCodes.js";
 
 const {
   serverErrors: { internalServerErrorCode },
   clientErrors: { notFoundCode },
 } = httpStatusCodes;
 
-const debug = debugCreator("api-gateway:middlewares:errors");
+const debug = debugCreator("identify-server:middlewares:errors");
 
 const generalError = (
   error: CustomError,
